refactor: make telemetry more lightweight and robust
