+++ jina/serve/executors/__init__.py
@@ -7,9 +7,15 @@ import warnings
 from types import SimpleNamespace
 from typing import TYPE_CHECKING, Any, Dict, Optional, Type, Union
 
-from jina import __args_executor_init__, __default_endpoint__
+from jina import __args_executor_init__, __cache_path__, __default_endpoint__
 from jina.enums import BetterEnum
-from jina.helper import ArgNamespace, T, iscoroutinefunction, typename
+from jina.helper import (
+    ArgNamespace,
+    T,
+    iscoroutinefunction,
+    send_telemetry_event,
+    typename,
+)
 from jina.importer import ImportExtensions
 from jina.jaml import JAML, JAMLCompatible, env_var_regex, internal_var_regex
 from jina.logging.logger import JinaLogger
@@ -145,6 +151,7 @@ class BaseExecutor(JAMLCompatible, metaclass=ExecutorType):
             )
         if type(self) == BaseExecutor:
             self.requests[__default_endpoint__] = self._dry_run_func
+        send_telemetry_event(event='start', obj=self)
 
     def _dry_run_func(self, *args, **kwargs):
         pass
@@ -325,7 +332,7 @@ class BaseExecutor(JAMLCompatible, metaclass=ExecutorType):
             getattr(self.runtime_args, 'workspace', None)
             or getattr(self.metas, 'workspace')
             or self._init_workspace
-            or os.environ.get('JINA_DEFAULT_WORKSPACE_BASE')
+            or __cache_path__
         )
         if workspace:
             complete_workspace = os.path.join(workspace, self.metas.name)
